server/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
python-dotenv==0.19.0
psycopg2-binary==2.9.1
alembic==1.7.1
tenacity==8.0.1
requests==2.26.0
python-dateutil==2.8.2
